requirements.txt
multidict==6.0.5
aiogram~=3.12.0
pyrogram~=2.0.106
mysql-connector-python~=9.0.0
aioconsole
python-dotenv